Fix for error at application-cec step
Fix for intermittent error, causing deployment to fail. Exception message, "System.Threading.ThreadAbortException: Thread was being aborted" or error message, "Deployment was interrupted and the process running it was terminated unexpectedly, if an ARM template is used, the likely cause is a race condition".
The fix involved porting back XC9.2 ARM template changes that addressed a similar issue. The primary fix was the introduction of more explicit "depends on" commands.
The changes also replaced two application-cec-as.json and application-cec-solr.json files with a single merged application-cec-image.json